Change backup/restore directories behaviour
Until now, when backuping/restoring directories,
we choosed to use this command:
rsync [...] /path/to/dir .../backups/.../path/to/

but in this case, if we use excludes command, context
is set to /path/to.
e.g. to ignore .cache directories, we had to put
/dir/.cache/ in exclude file.

So we changed it to use command:
rsync [...] /path/to/dir/ .../backups/.../path/to/dir/